# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  success: boolean;
  error: string;
  message?: string;
}

export interface VolumePrice {
  min_qty: number;
  price: number;
}

export interface SubscriptionType {
  slug: string;
  name: string;
  duration_days: number;
  price: number;
  volume_prices?: VolumePrice[];
  in_stock: boolean;
}

export interface Product {
  slug: string;
  name: string;
  subscription_types: SubscriptionType[];
}

export interface ProductsResponse {
  success: boolean;
  data: Product[];
}

export interface BalanceData {
  balance: number;
  currency: string;
}

export interface BalanceResponse {
  success: boolean;
  data: BalanceData;
}

export interface ValidateKeyBody {
  /** The CDK activation key to validate */
  key: string;
}

export type ValidateKeyResultStatus =
  (typeof ValidateKeyResultStatus)[keyof typeof ValidateKeyResultStatus];

export const ValidateKeyResultStatus = {
  valid: "valid",
  already_used: "already_used",
  invalid: "invalid",
} as const;

export interface ValidateKeyResult {
  status: ValidateKeyResultStatus;
  product?: string;
  subscription?: string;
  email?: string;
  activated_at?: string;
  message?: string;
}

export interface ValidateKeyResponse {
  success: boolean;
  data: ValidateKeyResult;
}

export interface ActivateKeyBody {
  /** The CDK activation key */
  key: string;
  /** Account access token */
  user_token: string;
  /** Use async activation mode */
  async?: boolean;
}

export interface ActivationData {
  email?: string;
  product?: string;
  subscription?: string;
  activated_at?: string;
  task_id?: string;
  status?: string;
  message?: string;
  warning?: string;
}

export interface ActivationResponse {
  success: boolean;
  data?: ActivationData;
  message?: string;
  warning?: string;
}

export interface OrderSummary {
  order_number: string;
  product: string;
  subscription: string;
  quantity: number;
  amount: number;
  status: string;
  created_at: string;
}

export interface OrdersMeta {
  current_page: number;
  last_page: number;
  per_page: number;
  total: number;
}

export interface OrdersResponse {
  success: boolean;
  data: OrderSummary[];
  meta?: OrdersMeta;
}

export interface OrderDetail {
  order_number: string;
  product: string;
  subscription: string;
  quantity: number;
  amount: number;
  status: string;
  created_at: string;
  keys?: string[];
  balance_after?: number;
  duration_days?: number;
}

export interface OrderDetailResponse {
  success: boolean;
  data: OrderDetail;
}

export interface PurchaseKeysBody {
  product_slug: string;
  subscription_type_slug: string;
  quantity?: number;
  idempotency_key?: string;
}

export interface PurchaseResponse {
  success: boolean;
  data?: OrderDetail;
  message?: string;
}

export type GetOrdersParams = {
  page?: number;
  per_page?: number;
};