# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Checks if a CDK key is valid, already used, or invalid without consuming it
 * @summary Validate a CDK key
 */
export const ValidateKeyBody = zod.object({
  key: zod.string().describe("The CDK activation key to validate"),
});

export const ValidateKeyResponse = zod.object({
  success: zod.boolean(),
  data: zod.object({
    status: zod.enum(["valid", "already_used", "invalid"]),
    product: zod.string().optional(),
    subscription: zod.string().optional(),
    email: zod.string().optional(),
    activated_at: zod.string().optional(),
    message: zod.string().optional(),
  }),
});

/**
 * Returns a list of available products with subscription types, prices, and stock status
 * @summary Get available products
 */
export const GetProductsResponse = zod.object({
  success: zod.boolean(),
  data: zod.array(
    zod.object({
      slug: zod.string(),
      name: zod.string(),
      subscription_types: zod.array(
        zod.object({
          slug: zod.string(),
          name: zod.string(),
          duration_days: zod.number(),
          price: zod.number(),
          volume_prices: zod
            .array(
              zod.object({
                min_qty: zod.number(),
                price: zod.number(),
              }),
            )
            .optional(),
          in_stock: zod.boolean(),
        }),
      ),
    }),
  ),
});

/**
 * Returns the current account balance
 * @summary Get account balance
 */
export const GetBalanceResponse = zod.object({
  success: zod.boolean(),
  data: zod.object({
    balance: zod.number(),
    currency: zod.string(),
  }),
});

/**
 * Activates a purchased key using the keys.ovh API
 * @summary Activate a CDK key
 */
export const ActivateKeyBody = zod.object({
  key: zod.string().describe("The CDK activation key"),
  user_token: zod.string().describe("Account access token"),
  async: zod.boolean().optional().describe("Use async activation mode"),
});

export const ActivateKeyResponse = zod.object({
  success: zod.boolean(),
  data: zod
    .object({
      email: zod.string().optional(),
      product: zod.string().optional(),
      subscription: zod.string().optional(),
      activated_at: zod.string().optional(),
      task_id: zod.string().optional(),
      status: zod.string().optional(),
      message: zod.string().optional(),
      warning: zod.string().optional(),
    })
    .optional(),
  message: zod.string().optional(),
  warning: zod.string().optional(),
});

/**
 * Returns a paginated list of API purchase orders
 * @summary Get purchase orders
 */
export const getOrdersQueryPageDefault = 1;
export const getOrdersQueryPerPageDefault = 20;

export const GetOrdersQueryParams = zod.object({
  page: zod.coerce.number().default(getOrdersQueryPageDefault),
  per_page: zod.coerce.number().default(getOrdersQueryPerPageDefault),
});

export const GetOrdersResponse = zod.object({
  success: zod.boolean(),
  data: zod.array(
    zod.object({
      order_number: zod.string(),
      product: zod.string(),
      subscription: zod.string(),
      quantity: zod.number(),
      amount: zod.number(),
      status: zod.string(),
      created_at: zod.string(),
    }),
  ),
  meta: zod
    .object({
      current_page: zod.number(),
      last_page: zod.number(),
      per_page: zod.number(),
      total: zod.number(),
    })
    .optional(),
});

/**
 * Returns details of a specific order including purchased keys
 * @summary Get a specific order
 */
export const GetOrderParams = zod.object({
  orderNumber: zod.coerce.string(),
});

export const GetOrderResponse = zod.object({
  success: zod.boolean(),
  data: zod.object({
    order_number: zod.string(),
    product: zod.string(),
    subscription: zod.string(),
    quantity: zod.number(),
    amount: zod.number(),
    status: zod.string(),
    created_at: zod.string(),
    keys: zod.array(zod.string()).optional(),
    balance_after: zod.number().optional(),
    duration_days: zod.number().optional(),
  }),
});

/**
 * Purchase activation keys using account balance
 * @summary Purchase activation keys
 */
export const purchaseKeysBodyQuantityDefault = 1;

export const PurchaseKeysBody = zod.object({
  product_slug: zod.string(),
  subscription_type_slug: zod.string(),
  quantity: zod.number().default(purchaseKeysBodyQuantityDefault),
  idempotency_key: zod.string().optional(),
});

export const PurchaseKeysResponse = zod.object({
  success: zod.boolean(),
  data: zod
    .object({
      order_number: zod.string(),
      product: zod.string(),
      subscription: zod.string(),
      quantity: zod.number(),
      amount: zod.number(),
      status: zod.string(),
      created_at: zod.string(),
      keys: zod.array(zod.string()).optional(),
      balance_after: zod.number().optional(),
      duration_days: zod.number().optional(),
    })
    .optional(),
  message: zod.string().optional(),
});