# lib/api-zod/src/generated/types/activateKeyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivateKeyBody {
  /** The CDK activation key */
  key: string;
  /** Account access token */
  user_token: string;
  /** Use async activation mode */
  async?: boolean;
}