# lib/api-zod/src/generated/types/activationData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivationData {
  email?: string;
  product?: string;
  subscription?: string;
  activated_at?: string;
  task_id?: string;
  status?: string;
  message?: string;
  warning?: string;
}