# lib/api-zod/src/generated/types/activationResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivationData } from "./activationData";

export interface ActivationResponse {
  success: boolean;
  data?: ActivationData;
  message?: string;
  warning?: string;
}