# lib/api-zod/src/generated/types/balanceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BalanceData } from "./balanceData";

export interface BalanceResponse {
  success: boolean;
  data: BalanceData;
}