# lib/api-zod/src/generated/types/orderDetailResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderDetail } from "./orderDetail";

export interface OrderDetailResponse {
  success: boolean;
  data: OrderDetail;
}