# lib/api-zod/src/generated/types/orderSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderSummary {
  order_number: string;
  product: string;
  subscription: string;
  quantity: number;
  amount: number;
  status: string;
  created_at: string;
}