# lib/api-zod/src/generated/types/ordersMeta.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrdersMeta {
  current_page: number;
  last_page: number;
  per_page: number;
  total: number;
}