# lib/api-zod/src/generated/types/ordersResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrdersMeta } from "./ordersMeta";
import type { OrderSummary } from "./orderSummary";

export interface OrdersResponse {
  success: boolean;
  data: OrderSummary[];
  meta?: OrdersMeta;
}