# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionType } from "./subscriptionType";

export interface Product {
  slug: string;
  name: string;
  subscription_types: SubscriptionType[];
}