# lib/api-zod/src/generated/types/purchaseKeysBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PurchaseKeysBody {
  product_slug: string;
  subscription_type_slug: string;
  quantity?: number;
  idempotency_key?: string;
}