# lib/api-zod/src/generated/types/subscriptionType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VolumePrice } from "./volumePrice";

export interface SubscriptionType {
  slug: string;
  name: string;
  duration_days: number;
  price: number;
  volume_prices?: VolumePrice[];
  in_stock: boolean;
}