# lib/api-zod/src/generated/types/validateKeyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ValidateKeyBody {
  /** The CDK activation key to validate */
  key: string;
}