# lib/api-zod/src/generated/types/validateKeyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ValidateKeyResult } from "./validateKeyResult";

export interface ValidateKeyResponse {
  success: boolean;
  data: ValidateKeyResult;
}