# lib/api-zod/src/generated/types/validateKeyResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ValidateKeyResultStatus } from "./validateKeyResultStatus";

export interface ValidateKeyResult {
  status: ValidateKeyResultStatus;
  product?: string;
  subscription?: string;
  email?: string;
  activated_at?: string;
  message?: string;
}